requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
pypdf>=4.0.0
Pillow>=10.0.0
fpdf2>=2.7.0